# Optimize performance of PostgreSQL timestamptz type