@kamipo 2018-12-17
The `DateTimeFast` is slower than the original if a value has no timezone offset.

```ruby
require 'benchmark/ips'

timestamps = ['2016-03-04 13:38:41.424845', '2012-10-03 10:57:03']

timestamps.each do |timestamp|
  Benchmark.ips do |x| 
    x.report("orig") {
      ActiveRecord::ConnectionAdapters::PostgreSQL::OID::DateTime.new.cast_value(timestamp)
    }

    x.report("new") {
      ActiveRecord::ConnectionAdapters::PostgreSQL::OID::DateTimeFast.new.cast_value(timestamp)
    }

    x.compare!
  end
end
```

```
Warming up --------------------------------------
                orig    12.770k i/100ms
                 new    11.750k i/100ms
Calculating -------------------------------------
                orig    133.963k (± 5.6%) i/s -    676.810k in   5.068026s
                new    121.833k (± 4.8%) i/s -    611.000k in   5.026673s

Comparison:
                orig:   133963.1 i/s
                 new:   121833.3 i/s - same-ish: difference falls within error

Warming up --------------------------------------
                orig    13.024k i/100ms
                 new    12.305k i/100ms
Calculating -------------------------------------
                orig    135.800k (± 7.0%) i/s -    690.272k in   5.109614s
                new    124.746k (± 6.4%) i/s -    627.555k in   5.055505s

Comparison:
                orig:   135799.5 i/s
                 new:   124746.1 i/s - same-ish: difference falls within error
```

How about extending the `ISO_DATETIME` itself to recognize timezone offset?