@stanhu 2018-12-17
Sure, that works:

```ruby
require 'active_record'
require 'active_support'
require 'active_record/connection_adapters/postgresql_adapter'
require 'benchmark/ips'

module ActiveRecord
  module ConnectionAdapters
    module PostgreSQL
      module OID # :nodoc:
        class DateTimeFast < DateTime # :nodoc:
         ISO_DATETIME_TZ = /\A(\d{4})-(\d\d)-(\d\d) (\d\d):(\d\d):(\d\d)(\.\d+)?([+|-]\d\d)?\z/

         private

         def fast_string_to_time(string)
           if string =~ ISO_DATETIME_TZ
             microsec = parse_microseconds($7)
             offset = $8.nil? ? nil : $8.to_i * 3600
             new_time $1.to_i, $2.to_i, $3.to_i, $4.to_i, $5.to_i, $6.to_i, microsec, offset
           end
         end
        end
      end
    end
  end
end

timestamps = ['2016-03-04 13:38:41.424845+00', '2012-10-03 10:57:03-02', '2012-10-03 10:57:03']

timestamps.each do |timestamp|
  Benchmark.ips do |x|
    x.report("orig") {
      ActiveRecord::ConnectionAdapters::PostgreSQL::OID::DateTime.new.send(:cast_value, timestamp)
    }

    x.report("new") {
      ActiveRecord::ConnectionAdapters::PostgreSQL::OID::DateTimeFast.new.send(:cast_value, timestamp)
    }

    x.compare!
  end
end
```

```
Warming up --------------------------------------
                orig     3.675k i/100ms
                 new    13.004k i/100ms
Calculating -------------------------------------
                orig     44.044k (± 2.2%) i/s -    220.500k in   5.008774s
                 new    137.927k (± 3.3%) i/s -    702.216k in   5.097371s

Comparison:
                 new:   137926.9 i/s
                orig:    44044.3 i/s - 3.13x  slower

Warming up --------------------------------------
                orig     4.514k i/100ms
                 new    13.268k i/100ms
Calculating -------------------------------------
                orig     48.028k (± 2.0%) i/s -    243.756k in   5.077280s
                 new    139.670k (± 2.4%) i/s -    703.204k in   5.037825s

Comparison:
                 new:   139670.3 i/s
                orig:    48027.7 i/s - 2.91x  slower

Warming up --------------------------------------
                orig    13.772k i/100ms
                 new    14.767k i/100ms
Calculating -------------------------------------
                orig    159.514k (± 1.8%) i/s -    798.776k in   5.009177s
                 new    156.922k (± 3.4%) i/s -    797.418k in   5.087759s

Comparison:
                orig:   159514.5 i/s
                 new:   156922.3 i/s - same-ish: difference falls within error
```