@kamipo 2019-02-27
Can you try to use master or beta2?
I think that the performance of timestamptz from database will be improved by #35062.